+++ database/migrations/2019_11_06_151203_create_reply_rule_tables.php
@@ -14,7 +14,7 @@ class CreateReplyRuleTables extends Migration
     public function up()
     {
         Schema::create('we_rule', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('app_id', 20);
             $table->string('scene', 9);
             $table->string('title', 128);
@@ -28,8 +28,8 @@ class CreateReplyRuleTables extends Migration
         });
 
         Schema::create('we_keyword', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('rule_id');
+            $table->increments('id');
+            $table->unsignedInteger('rule_id');
             $table->string('keyword', 64);
             $table->unsignedTinyInteger('match_type')->comment('匹配规则: 1全匹配 2半匹配');
             $table->timestamps();
@@ -43,8 +43,8 @@ class CreateReplyRuleTables extends Migration
         });
 
         Schema::create('we_reply', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('rule_id');
+            $table->increments('id');
+            $table->unsignedInteger('rule_id');
             $table->unsignedTinyInteger('difference')->comment('是否区分男女: 0否 1是');
             $table->unsignedTinyInteger('reply_type')->comment('回复消息类型: 1文本 2图文 3图片 4音频 5视频 6位置 7链接 8小程序 9任务宝 10拼团 11分销 12优惠券');
             $table->unsignedTinyInteger('reply_type_female');
